html as a family tree
See notes in code; it compares the different parts of a html code as
parents, siblings and children.
If you think of the <html> tag as the trunk of the tree, you can think
of its immediate branches—<head> and <body>—as its children. Both tags
are children of <html>, and <html> is their parent element. Because
they are both immediate children of <html> (that is, they are both only
one element away), they are siblings.

Just like a real family, elements have children, grandchildren,
great-grandchildren, and so on (though we don't make this distinction
with HTML—a child of an element, and all that child's children, are
children of the first parent).
Remember: an element is a child of EVERY element wrapped around it,
even if that element is several "branches" away!