Classes
Classes are useful when you have a bunch of elements that should all
receive the same styling.
So rather than applying the same rules to different selectors, you can
simply apply toe same class to all those html elements, then define the
styling for that class in the css doc.
Classes are assigned to a html doc with the word class and equals sign.
You can make the word any word you like for the class identifier e.g.
<div class=“this_one”></div>