id selector
Id selector is great for when you have exactly one element that should
receive on kind of styling.
In html, id’s are assigned with id equal to an identifying word within
the element.
In css, the identifying word is used with the hashtag or pound # symbol
in the front.
This allows you to apply a style to a single instance of a selector,
rather than all instances.