Display inline
The inline-block value allows you to put several block elements on the
same line. The inline value places all your elements next to one
another, but not as blocks: they don't keep their dimensions.
Note: the displayed part of the div’s word does have a small box around
it, but is only the size that the text is. So the word two’s box is
shorter than fours.