padding
See notes in css code
Note: the padding just made the colour of the block get bigger, as the
background colour is the same for the content and the padding.
When you give CSS a positive padding or margin value, it puts that
space between the element and its reference: for instance, if you have
a <div> and you give it a margin-left of 20px, it puts twenty pixels
between the left margin of that <div> and the side of the screen. This
effectively moves the <div> twenty pixels to the right.