Clear
If the floating elements mix with non floating elements (start to
overlap), you can tell an element to clear: left, right or both and the
element will move below any floating elements. e.g. clear:left means
the element will move below any floating elements on the left side of
the page, clear:right will do the same on the right, and clear:both
will get out of the way of elements floating on the left and the right.