Absolute positioning
If you don't specify an element's positioning type, it defaults to
static. This just means "where the element would normally go." If you
don't tell an element how to position itself, it just plunks itself
down in the document.
The first type of positioning is absolute positioning. When an element
is set to position: absolute, it's then positioned in relation to the
first parent element it has that doesn't have position: static. If
there's no such element, the element with position: absolute gets
positioned relative to <html>.

To show you how this works, we've set the #outer div to have absolute
positioning. This means that when you position the #inner div, it will
be relative to #outer. (If #outer had the default positioning of
static, then #inner would get positioned relative to the entire HTML
document.)