Resume part 3
Adding in a position:fixes and z-index:1
 You can think of the z-index as a measure of importance: the higher an
element's z-index, the higher it will be "stacked" on the page. Giving
your header a z-index of 1 while not giving any of the other elements a
z-index ensures that your header will sit "on top of" your other
elements and won't get stuck behind them.